When trying to deploy on non-linode systems, there are a few issues.
1. The font location, names, and even number, can change.  Put this
into the lacuna.conf.template, but default to linode's values.

2. Setting up beanstalk - we can put this right into the template
(and can be commented out if not needed, like much of the rest of the
template)

3. A few errors during execution: medium_blob won't deploy, and a ++ found
somewhere.

With all of this, I can get a running system up a fair bit faster,
though some documentation on the captcha bit is likely required.