An attempt to get rid of the wild resource (including happy) fluctuations.
The planet tick would create a single query to return all occupied bodies,
but the sql server would return the state of all the rows at the *beginning*
of the query, so if those rows were changed during the long-running tick
process, it would be based on the resources from the beginning of the tick,
and thus lose all the changes done during the tick.

This can include building (using resources, generating/losing happy) or
spy actions (rebellions, etc.).

The code here will not pull full objects down, but only the ID and then
use separate queries to pull down the objects, which will hopefully
be in separate transactions and thus be as current as possible at the time
that body's tick is processed.