Massive cleanup.
Replace the get_body/get_building APIs in Lacuna::RPC with a single
get_session.  Expand Lacuna::Session to track current body/building/empire.

The primary purpose of this is to allow all these to be grabbed in
a single call (which is not part of this commit), reducing the DB hits.
And, more specifically, to do this for the sitter changes coming up.
This part is delivered separately so that it can be exercised now while
the sitter changes continue to be worked on.

I'm also reducing some red tape here for the most trivial of module
actions (repair, demolish, downgrade).  This won't eliminate much excess
voting email, but it's a start.  Upgrade and build are the two biggest ones,
but they involve figuring out the appropriate time to attach to the build
actions before removing them, and they impact more than just the immediate
code in Building.pm.