Modify the sitter auth cleanup.
We don't delete expired auths, but, instead, only delete them when one
of the two named empires are deleted.  This means that if your sole
sitter gets deleted, it won't be counted.

Then, the empire cleanup code will look if there are any sitter auths
that expired recently, or have not yet expired.  If so, we just skip
that empire altogether, it won't get cleaned up.  This should extend
an empire out to 20 days after the last sitter expires, or 20 days
after its last login, whichever comes _later_.  And then the regular
essentia drain kicks in.  While we'd like to change that as well,
that change is not part of this update.