Add extra info to the trade buildings for pushes.
TM: the list should include all planets and stations can be pushed to,
even for unallied babies (this way a sitter can still push to stations).

SST: the list should include only the planets that have an online SST
to receive with.

The views look the same to the UI to make them easier to implement.